+++ setup.cfg
@@ -10,4 +10,4 @@ exclude = env,build
 [nosetests]
 cover-branches = 1
 cover-erase = 1
-cover-package = sprockets.clients.influxdb
+cover-package = sprockets_influxdb
